% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_dcc.R
\name{anova_dcc}
\alias{anova_dcc}
\title{ANOVA do modelo DCC}
\usage{
anova_dcc(obj, digits = 4)
}
\arguments{
\item{obj}{objeto da classe dcc_fit.}

\item{digits}{número de casas decimais.}
}
\value{
tabela ANOVA formatada.
}
\description{
ANOVA do modelo DCC
}
