% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficientes_dcc.R
\name{coeficientes_dcc}
\alias{coeficientes_dcc}
\title{Tabela formatada de coeficientes do modelo DCC}
\usage{
coeficientes_dcc(obj, digits = 4)
}
\arguments{
\item{obj}{objeto da classe dcc_fit.}

\item{digits}{número de casas decimais.}
}
\value{
tabela formatada.
}
\description{
Tabela formatada de coeficientes do modelo DCC
}
