% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcc_terms.R
\name{dcc_terms}
\alias{dcc_terms}
\title{Gerar termos quadráticos e de interação para DCC}
\usage{
dcc_terms(dados, fatores)
}
\arguments{
\item{dados}{data.frame com os dados experimentais.}

\item{fatores}{vetor com os nomes dos fatores.}
}
\value{
data.frame organizado com AA, BB e AB antes da resposta.
}
\description{
Gerar termos quadráticos e de interação para DCC
}
