% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efeitos_dcc.R
\name{efeitos_dcc}
\alias{efeitos_dcc}
\title{Efeitos estimados e significância no DCC}
\usage{
efeitos_dcc(obj, alpha = 0.05, digits = 4)
}
\arguments{
\item{obj}{objeto da classe dcc_fit.}

\item{alpha}{nível de significância.}

\item{digits}{casas decimais.}
}
\value{
tabela de efeitos com significância.
}
\description{
Efeitos estimados e significância no DCC
}
