% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_matriz_dcc.R
\name{export_matriz_dcc}
\alias{export_matriz_dcc}
\title{Exporta a matriz experimental DCC}
\usage{
export_matriz_dcc(matriz, arquivo = "Matriz_DCC", remover_ensaio = FALSE)
}
\arguments{
\item{matriz}{data.frame gerado pela funcao matriz_dcc().}

\item{arquivo}{nome do arquivo de saida, sem extensao.}

\item{remover_ensaio}{logical. Se TRUE, remove a coluna Ensaio antes da exportacao.}
}
\value{
Invisivelmente, o nome do arquivo gerado.
}
\description{
Exporta a matriz experimental DCC
}
