% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_relatorio_dcc.R
\name{exportar_relatorio_dcc}
\alias{exportar_relatorio_dcc}
\title{Exporta relatório completo do DCC}
\usage{
exportar_relatorio_dcc(
  fit,
  arquivo = "Relatorio_DCC",
  formato = "txt",
  tipo_otimo = NULL
)
}
\arguments{
\item{fit}{objeto da classe dcc_fit}

\item{arquivo}{nome do arquivo a ser salvo, sem extensão}

\item{formato}{"txt" ou "doc"}

\item{tipo_otimo}{NULL para não incluir ótimo, "min" para minimizar
a resposta ou "max" para maximizar}
}
\value{
invisivelmente, o nome do arquivo gerado
}
\description{
Exporta relatório completo do DCC
}
