% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_dcc.R
\name{matriz_dcc}
\alias{matriz_dcc}
\title{Gera a matriz do Delineamento Composto Central}
\usage{
matriz_dcc(
  k = 2,
  alpha = "rotacional",
  aleatorizar = FALSE,
  seed = NULL,
  nome_resposta = "Resposta",
  incluir_tipo_ponto = TRUE
)
}
\arguments{
\item{k}{número de fatores, de 2 a 5.}

\item{alpha}{valor de alpha ou "rotacional".}

\item{aleatorizar}{lógico. Se TRUE, embaralha a ordem dos ensaios.}

\item{seed}{valor opcional para reprodutibilidade da aleatorização.}

\item{nome_resposta}{nome da coluna de resposta.}

\item{incluir_tipo_ponto}{lógico. Se TRUE, inclui a coluna Tipo_ponto.}
}
\value{
data.frame com a matriz do DCC.
}
\description{
Gera a matriz do Delineamento Composto Central
}
