% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricas_dcc.R
\name{metricas_dcc}
\alias{metricas_dcc}
\title{Métricas do modelo DCC}
\usage{
metricas_dcc(obj, digits = 4)
}
\arguments{
\item{obj}{objeto da classe dcc_fit.}

\item{digits}{número de casas decimais.}
}
\value{
tabela formatada com métricas do modelo.
}
\description{
Métricas do modelo DCC
}
