% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opcoes_dcc.R
\name{opcoes_dcc}
\alias{opcoes_dcc}
\title{Definir opcoes de exibicao do pacote DCC}
\usage{
opcoes_dcc(decimal = c("virgula", "ponto"))
}
\arguments{
\item{decimal}{Marca decimal a ser usada. Pode ser "virgula" ou "ponto".}
}
\value{
Invisivelmente, a marca decimal ativa.
}
\description{
Controla a marca decimal usada na exibicao dos resultados.
}
