% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimo_dcc.R
\name{otimo_dcc}
\alias{otimo_dcc}
\title{Ponto ótimo previsto para DCC}
\usage{
otimo_dcc(fit, objetivo = c("min", "max"))
}
\arguments{
\item{fit}{objeto da classe dcc_fit.}

\item{objetivo}{"min" para minimizar a resposta ou "max" para maximizar.}
}
\value{
Lista com o ponto ótimo previsto, resposta estimada,
convergência, valor otimizado e objetivo.
}
\description{
Ponto ótimo previsto para DCC
}
