% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_dcc.R
\name{pareto_dcc}
\alias{pareto_dcc}
\title{Gráfico de Pareto dos efeitos padronizados no DCC}
\usage{
pareto_dcc(obj, alpha = 0.05)
}
\arguments{
\item{obj}{objeto da classe dcc_fit.}

\item{alpha}{nível de significância.}
}
\value{
Invisivelmente, a tabela usada no gráfico.
}
\description{
Gráfico de Pareto dos efeitos padronizados no DCC
}
