% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dcc_fit.R
\name{print.dcc_fit}
\alias{print.dcc_fit}
\title{Imprimir objeto ajustado do tipo dcc_fit}
\usage{
\method{print}{dcc_fit}(x, ...)
}
\arguments{
\item{x}{objeto da classe dcc_fit.}

\item{...}{argumentos adicionais.}
}
\value{
Invisivelmente, o próprio objeto.
}
\description{
Imprimir objeto ajustado do tipo dcc_fit
}
