% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumario_dcc.R
\name{sumario_dcc}
\alias{sumario_dcc}
\title{Sumario geral da analise DCC}
\usage{
sumario_dcc(fit, incluir_otimo = TRUE, imprimir = TRUE)
}
\arguments{
\item{fit}{objeto da classe dcc_fit.}

\item{incluir_otimo}{logical. Se TRUE, tenta incluir o ponto otimo.}

\item{imprimir}{logical. Se TRUE, imprime os resultados no console.}
}
\value{
Invisivelmente, uma lista com os resultados da analise.
}
\description{
Reune em uma unica funcao os principais resultados nao graficos
da analise de um Delineamento Composto Central.
}
