% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contorno_dcc.R
\name{contorno_dcc}
\alias{contorno_dcc}
\title{Gráfico de contorno para DCC}
\usage{
contorno_dcc(
  fit,
  x1,
  x2,
  n = 100,
  mostrar_otimo = TRUE,
  mostrar_estacionario = TRUE,
  objetivo = "max"
)
}
\arguments{
\item{fit}{objeto da classe dcc_fit.}

\item{x1}{nome do primeiro fator.}

\item{x2}{nome do segundo fator.}

\item{n}{número de pontos da grade.}

\item{mostrar_otimo}{se TRUE, exibe o ponto ótimo no gráfico.}

\item{mostrar_estacionario}{se TRUE, exibe o ponto estacionário no gráfico.}

\item{objetivo}{objetivo da otimização: "max" para maximizar ou "min" para minimizar.}
}
\value{
Gráfico de contorno.
}
\description{
Gráfico de contorno para DCC
}
