% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otimo_dcc.R
\name{otimo_dcc}
\alias{otimo_dcc}
\title{Ponto ótimo previsto para DCC}
\usage{
otimo_dcc(fit, tipo = c("min", "max"))
}
\arguments{
\item{fit}{objeto da classe dcc_fit.}

\item{tipo}{"min" para minimizar a resposta ou "max" para maximizar.}
}
\value{
data.frame com o ponto ótimo previsto e o valor estimado da resposta.
}
\description{
Ponto ótimo previsto para DCC
}
