% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clipboard_dcc.R
\name{read_clipboard_dcc}
\alias{read_clipboard_dcc}
\title{Ler dados copiados da planilha para DCC}
\usage{
read_clipboard_dcc()
}
\value{
Um data.frame com os dados lidos do clipboard.
}
\description{
Lê do clipboard uma tabela copiada do Excel, esperando colunas
com os fatores codificados e a resposta.
}
