% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superficie_dcc.R
\name{superficie_dcc}
\alias{superficie_dcc}
\title{Gráfico de superfície de resposta para DCC}
\usage{
superficie_dcc(fit, x1, x2, n = 45)
}
\arguments{
\item{fit}{objeto da classe dcc_fit.}

\item{x1}{nome do primeiro fator.}

\item{x2}{nome do segundo fator.}

\item{n}{número de pontos da grade.}
}
\value{
Gráfico de superfície.
}
\description{
Gráfico de superfície de resposta para DCC
}
